MODOAIPMH-626 - Upgrade module to Vert.x 5.0

+++ src/test/java/org/folio/oaipmh/service/impl/SetServiceImplTest.java
@@ -35,7 +35,6 @@ import org.folio.oaipmh.dao.PostgresClientFactory;
 import org.folio.oaipmh.dao.SetDao;
 import org.folio.oaipmh.dao.impl.SetDaoImpl;
 import org.folio.oaipmh.service.SetService;
-import org.folio.okapi.common.GenericCompositeFuture;
 import org.folio.postgres.testing.PostgresTesterContainer;
 import org.folio.rest.impl.OkapiMockServer;
 import org.folio.rest.jaxrs.model.FolioSet;
@@ -82,7 +81,7 @@ class SetServiceImplTest extends AbstractSetTest {
   void tearDownClass(Vertx vertx, VertxTestContext testContext) {
     PostgresClientFactory.closeAll();
     WebClientProvider.closeAll();
-    vertx.close(testContext.succeeding(res -> {
+    vertx.close().onComplete(testContext.succeeding(res -> {
       testContext.completeNow();
     }));
   }
@@ -274,13 +273,15 @@ class SetServiceImplTest extends AbstractSetTest {
     setDao.getSetList(0, 100, OAI_TEST_TENANT)
         .onComplete(result -> {
           FolioSetCollection setItemCollection = result.result();
-          List<Future> futures = new ArrayList<>();
-          setItemCollection.getSets()
-              .forEach(setItem -> futures.add(setDao.deleteSetById(setItem.getId(),
-                  OAI_TEST_TENANT)));
-          GenericCompositeFuture.all(futures)
-              .onSuccess(compositeFuture -> testContext.completeNow())
-              .onFailure(testContext::failNow);
+          List<Future<Boolean>> futures = new ArrayList<>();
+          if (setItemCollection != null) {
+            setItemCollection.getSets()
+                .forEach(setItem -> futures.add(setDao.deleteSetById(setItem.getId(),
+                    OAI_TEST_TENANT)));
+            Future.all(futures)
+                .onSuccess(compositeFuture -> testContext.completeNow())
+                .onFailure(testContext::failNow);
+          }
         });
   }
